05_streamlit/requirements.txt
openai==1.17.1
pandas==2.2.2
PyYAML==6.0.1
streamlit==1.33.0